refactor: move bindings
